+++ ketraterm-intellij-plugin/CHANGELOG.md
@@ -1,5 +1,9 @@
 # KetraTerm IntelliJ Plugin Changelog
 
+## [Unreleased]
+
+- Fixed first-run TUI wrapping corruption.
+
 ## [0.1.1] - 2026-07-02
 
 - Fixed AltGr character input and terminal grid corruption on AZERTY keyboards for Windows/Linux.